pyproject.toml
[tool.poetry]
name = "Show"
version = "0.0.2"
description = "Shaders On Wallpaper, a live and interactive wallpaper 'engine' for linux"
authors = [ "DanielFvM <ds@xxxmail.eu>" ]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/danielfvm/show"
repository = "https://github.com/danielfvm/show"
keywords = [ "linux", "shaders", "wallpapers", "wayland", "gif-animation", "glsl-shaders", "wallpaper-engine" ]
packages = [ 
	{ include = "show" }
]

[tool.poetry.urls]
"Homepage" = "https://github.com/danielfvm/show"
"Bug Tracker" = "https://github.com/danielfvm/show/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
xcffib = "^0.11.1"
cairocffi = "^1.3.0"
pyopengl = "^3.1.6"
screeninfo = "^0.8"
mouse = "^0.7.1"
glfw = "^2.5.3"
Pillow = "^9.2.0"

[tool.poetry.scripts]
show = "show.main:main"
